create EnumUtils plus isDevEnvironment
