+++ .idea/dictionaries/default_user.xml
@@ -126,6 +126,7 @@
       <w>jawbus</w>
       <w>jerries</w>
       <w>jerrypocalypse</w>
+      <w>jsonobjects</w>
       <w>jyrre</w>
       <w>kada</w>
       <w>kaeso</w>
@@ -241,6 +242,7 @@
       <w>skyblock</w>
       <w>skyhanni</w>
       <w>skyhanni's</w>
+      <w>skyhannimodule</w>
       <w>skymall</w>
       <w>skytils</w>
       <w>skytils's</w>
