+++ src/main/java/at/hannibal2/skyhanni/api/ReforgeApi.kt
@@ -7,12 +7,12 @@ import at.hannibal2.skyhanni.data.model.SkyblockStatList
 import at.hannibal2.skyhanni.events.NeuRepositoryReloadEvent
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.test.command.ErrorManager
+import at.hannibal2.skyhanni.utils.EnumUtils
 import at.hannibal2.skyhanni.utils.ItemCategory
 import at.hannibal2.skyhanni.utils.ItemUtils.getInternalName
 import at.hannibal2.skyhanni.utils.ItemUtils.getItemCategoryOrNull
 import at.hannibal2.skyhanni.utils.ItemUtils.itemNameWithoutColor
 import at.hannibal2.skyhanni.utils.LorenzRarity
-import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.NeuInternalName
 import at.hannibal2.skyhanni.utils.json.BaseGsonBuilder
 import at.hannibal2.skyhanni.utils.json.SkyHanniTypeAdapters
@@ -200,7 +200,7 @@ object ReforgeApi {
         val type = it.itemType
         return Reforge(
             name = it.reforgeName,
-            type = LorenzUtils.enumValueOf<ReforgeType>(type.first),
+            type = EnumUtils.enumValueOf<ReforgeType>(type.first),
             stats = it.reforgeStats.orEmpty(),
             reforgeStone = it.internalName,
             specialItems = type.second.takeIf { it.isNotEmpty() },
