+++ src/main/java/at/hannibal2/skyhanni/events/WidgetUpdateEvent.kt
@@ -2,7 +2,7 @@ package at.hannibal2.skyhanni.events
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
 import at.hannibal2.skyhanni.data.model.TabWidget
-import at.hannibal2.skyhanni.utils.LorenzUtils.isAnyOf
+import at.hannibal2.skyhanni.utils.EnumUtils.isAnyOf
 
 /** The events get send on change of the widget and on island switch */
 open class WidgetUpdateEvent(
