+++ src/main/java/at/hannibal2/skyhanni/features/misc/items/GlowingDroppedItems.kt
@@ -6,6 +6,7 @@ import at.hannibal2.skyhanni.data.IslandType
 import at.hannibal2.skyhanni.events.RenderEntityOutlineEvent
 import at.hannibal2.skyhanni.features.garden.pests.SprayType
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
+import at.hannibal2.skyhanni.utils.EnumUtils
 import at.hannibal2.skyhanni.utils.ItemUtils.getInternalNameOrNull
 import at.hannibal2.skyhanni.utils.ItemUtils.getItemRarityOrNull
 import at.hannibal2.skyhanni.utils.LorenzUtils
@@ -59,7 +60,7 @@ object GlowingDroppedItems {
         }
 
         val internalName = entityItem.getInternalNameOrNull() ?: return null
-        val isSprayItem = LorenzUtils.enumValueOfOrNull<SprayType>(internalName.asString()) != null
+        val isSprayItem = EnumUtils.enumValueOfOrNull<SprayType>(internalName.asString()) != null
         if (isSprayItem) return null
         val rarity = entityItem.getItemRarityOrNull()
         return rarity?.color?.toColor()?.rgb
