+++ src/main/java/at/hannibal2/skyhanni/utils/ItemUtils.kt
@@ -392,7 +392,7 @@ object ItemUtils {
             UtilsPatterns.sackPattern.matches(name) -> ItemCategory.SACK
             else -> ItemCategory.NONE
         } else {
-            LorenzUtils.enumValueOfOrNull<ItemCategory>(itemCategory)
+            EnumUtils.enumValueOfOrNull<ItemCategory>(itemCategory)
         }
 
     private fun ItemStack.updateCategoryAndRarity() {
